new p2p proc file
